Add AppController centralizing gaze control
